+++ src/App.tsx
@@ -3,10 +3,7 @@ import { Box, useTheme } from "@mui/material";
 import { AuthProvider } from "@store/contexts/AuthContext";
 import { Topbar } from "@components/Topbar";
 import { MainLayout } from "@layouts/MainLayout";
-import { AuthLayout } from "@layouts/AuthLayout";
 import { HomePage } from "@pages/HomePage";
-import { LoginPage } from "@pages/LoginPage";
-import { SignUpPage } from "@pages/SignUpPage";
 import { SetupPage } from "@pages/SetupPage";
 import { AuthCallbackPage } from "@pages/AuthCallbackPage";
 
@@ -24,10 +21,6 @@ function AppContent() {
         <Routes>
           <Route element={<MainLayout />}>
             <Route path="/" element={<HomePage />} />
-          </Route>
-          <Route element={<AuthLayout />}>
-            <Route path="/login" element={<LoginPage />} />
-            <Route path="/signup" element={<SignUpPage />} />
           </Route>
           <Route path="/auth/callback" element={<AuthCallbackPage />} />
           <Route path="/setup" element={<SetupPage />} />