+++ src/pages/SignUpPage.tsx
@@ -1,5 +1,0 @@
-import { SignUpForm } from "@features/auth/components/SignUpForm";
-
-export const SignUpPage = () => {
-  return <SignUpForm />;
-};